Tests: move contents of data/testall to test_data and refactor it into a class
